requirements.txt
Flask==3.1.2
Flask-Bcrypt==1.0.1
python-dotenv==1.1.1
mysql-connector-python==9.1.0